all routes are now done by react and not with hrefs.
this improves a lot the time to open new pages
