+++ frontend/src/App/App.tsx
@@ -1,21 +1,18 @@
-import { AppBar, Box, IconButton, Toolbar } from "@material-ui/core";
+import { Box } from "@material-ui/core";
 import Container from "@material-ui/core/Container";
-import DashboardIcon from '@material-ui/icons/Dashboard';
 import React, { useState } from "react";
 import { BrowserRouter as Router, Route } from "react-router-dom";
 import Challenge from "../Challenge/Challenge";
 import Challenges, { ChallengeData } from "../Challenge/Challenges";
 import Dashboard from "../dashboard/Dashboard";
 import Documentation from "../Documentation/Documentation";
-import LogoMenu from '../Images/logo_menu.png';
+import Appbar from "./Appbar";
 import { LayoutContext } from "./LayoutContext";
-import Progress from "./Progress";
 import { ProgressService } from "./ProgressService";
-import useStyles from "./styles";
+
 
 const App = () => {
 
-  const classes = useStyles();
 
   const p = ProgressService.createOrGet();
 
@@ -37,25 +34,12 @@ const App = () => {
   };
 
 
-
   return (
-
-
     <div className="App">
-      <LayoutContext.Provider value={layoutInitialState}>
-        <AppBar position="sticky" className={classes.appbar}>
-          <Toolbar className={classes.toolbar}>
-            <Box className={classes.menuLeft}>
-              <IconButton edge="start" href={"/"} className={classes.menuButton} color="inherit" aria-label="menu">
-                <DashboardIcon />
-              </IconButton>
-              <img className={classes.menuLogo} src={LogoMenu} />
-            </Box>
-            <Progress />
-          </Toolbar>
-        </AppBar>
-        <Container fixed>
-          <Router>
+      <Router>
+        <LayoutContext.Provider value={layoutInitialState}>
+          <Appbar />
+          <Container fixed>
             {
               Challenges.map((c) => {
                 return (
@@ -67,9 +51,9 @@ const App = () => {
             }
             <Route exact path="/"><Dashboard /></Route>
             <Route exact path="/docs/:topic" component={Documentation} />
-          </Router>
-        </Container>
-      </LayoutContext.Provider>
+          </Container>
+        </LayoutContext.Provider>
+      </Router>
     </div >
   );
 };
