+++ frontend/src/ChallengeCard/ChallengeCard.tsx
@@ -1,11 +1,11 @@
-import { Chip } from "@material-ui/core";
-import Button from "@material-ui/core/Button";
+import { Button, Chip } from "@material-ui/core";
 import Card from "@material-ui/core/Card";
 import CardActionArea from "@material-ui/core/CardActionArea";
 import CardActions from "@material-ui/core/CardActions";
 import CardContent from "@material-ui/core/CardContent";
 import Typography from "@material-ui/core/Typography";
 import React from "react";
+import { useHistory } from "react-router";
 import { IChallengeCardProps } from "./IChallengeCardProp";
 
 
@@ -16,16 +16,22 @@ const pad = (n: number, size = 2): string => {
   return s;
 };
 
-
 const ChallengeCard = (props: IChallengeCardProps) => {
 
+  const history = useHistory();
   const done = <Chip size="small" label="Done" style={{ backgroundColor: 'green', color: 'white' }} />;
+
+  const onChallengeClick = (path: string) => {
+    return () => {
+      history.push(path);
+    };
+  };
 
 
   return (
     <Card>
       <CardActionArea>
-        <CardContent style={{ height: '180px' }}>
+        <CardContent style={{ height: '180px' }} onClick={onChallengeClick(props.challenge.url)}>
           <Typography gutterBottom variant="h5" component="h2" style={{ height: '40px' }}>
             {pad(props.index + 1, 2)} - {props.challenge.name}
           </Typography>
@@ -35,11 +41,10 @@ const ChallengeCard = (props: IChallengeCardProps) => {
         </CardContent>
       </CardActionArea>
       <CardActions>
-        <Button size="small" color="primary" href={props.challenge.url}>
+        <Button size="small" color="primary" onClick={onChallengeClick(props.challenge.url)}>
           Take the challenge
         </Button>
         {props.done ? done : ""}
-
       </CardActions>
     </Card >
   );
